refactor: organize navbar into ui package
