+++ components/ui/__init__.py
@@ -1,0 +1,5 @@
+"""Shared UI components."""
+
+from .navbar import create_navbar_layout, register_navbar_callbacks
+
+__all__ = ["create_navbar_layout", "register_navbar_callbacks"]
